Extract proposal lifecycle query boundaries
